factoring out some code into modules
